Win32: Change the behaviour of max_retries and retry_delay
There was an intermittent failure in the CustomAllocatorsTest
suite on Windows. This was caused by setting the input
options to std_mock, which used default values of zero
for max_retries and retry_delay.

Since having zero retries is somewhat of a special case,
don't default to max_retries/retry_delay=0 in the
options equating to zero retries/delay. Instead indicate
this by explicitly setting a negative value for those options.